requirements.txt
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.15
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7
django-celery==3.1.16
django-rest-swagger==0.3.4
django-toolbelt==0.0.1
djangorestframework==3.2.3
djangorestframework-gis==0.9.4
gunicorn==19.3.0
kombu==3.0.26
Pillow==2.9.0
psycopg2==2.6.1
pytz==2015.6
PyYAML==3.11
requests==2.7.0
static3==0.6.1
wheel==0.24.0
